backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
gunicorn==23.0.0
MarkupSafe==3.0.2
numpy==2.3.2
pillow==11.3.0
Werkzeug==3.1.3
rembg==2.0.56
